requirements.txt
anyio==4.13.0
asgiref==3.11.1
black==26.1.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
django-schema-viewer==0.5.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mypy_extensions==1.1.0
packaging==26.1
pathspec==1.0.3
platformdirs==4.5.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.2
pytokens==0.4.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
uuid6==2025.0.1
whitenoise==6.12.0
